Updated the newAccount Updater to handle multiple run without failing
